Planet proportions are now accurate to real size
The position and other parameter are not set just yet.
I plan on making the POV to navigate between planets before implementing the positions.
The reason for that being: with how gigantic the space between planets are and how small the actual planets are compared to the former, it'll be hard to find them on the display.