Changed implementation of event handling of camera
We now have a Context class that will contain all the objects needed in the event management. Objects such as the Camera or the planets position data.

We now associate a Context object to the window via glfwSetWindowUserPointer function.

We then recover the Context object in the callback function with glfwGetWindowUserPointer.

Note that glfwGetWindowUserPointer returns a void* and needs to be static casted everytime.
It might be interesting to set a static context field to the Events class to avoid having to cast everytime.