Added all other planets + some other changes
All the planets of the solar system have been added.

Some texture have been changed for better ones.

Decreased the rotation speed of planets by decreasing the rotation unit from 24 to 6.
This is because some planets like Jupiter rotates way too fast with a rotation unit of 24.
As a side note, it means that the revolution period is multiplied by 4.