Added torus class + ring of saturn
A Torus class has been created to render toruses.
PlanetData and planetObject have gained a few more fields to describe the ring.
I added an arbitrary ring around Saturn as a test.
The texture is not applied well on the Torus however.

TODO: solve the texture application on the torus.