Rings now have accurate proportions
Modified torus creation method: Since scaling a torus will change both the torus radius and pipe radius, I created individual toruses for each ring.
The creation of the toruses is done by the function createPlanetRing of the renderEngine.
The ID of the vao and vbo of the toruses are stored in vectors in the render engine.
Size conversions for the rings are done.